+++ src/backend/adapters/base-property.ts
@@ -14,7 +14,9 @@ const TITLE_COLUMN_NAMES = ['title', 'name', 'subject', 'email']
  * @property {string} datetime    date with time
  * @property {string} mixed       type representing an object
  * @property {string} reference   many to one reference
+ * @property {string} richtext    wysiwig editor
  * @property {string} textarea    resizable textarea input
+ * @property {string} password    password field
  */
 export type PropertyType =
   'string' | 'float' | 'number' | 'boolean' |